% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boards.R
\name{local_directory}
\alias{local_directory}
\alias{set_local_boards_directory}
\alias{get_local_boards_directory}
\title{Use local boards directory}
\usage{
set_local_boards_directory(dir)

get_local_boards_directory()
}
\arguments{
\item{dir}{Path to the selected location}
}
\description{
Use local boards directory
}
\examples{
\dontrun{
  set_local_boards_directory(tempdir())
  get_local_boards_directory()
}
}
