% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pins.R
\name{manage-pins}
\alias{manage-pins}
\alias{show_pins}
\alias{get_pins}
\title{Manage pins}
\usage{
show_pins(project = NULL)

get_pins(project = NULL)
}
\arguments{
\item{project}{Project name}
}
\description{
Manage pins
}
\section{Functions}{
\itemize{
\item \code{show_pins()}: Show all the pins or all pins in given project

\item \code{get_pins()}: Get all the pins or all pins in given project

}}
