% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{projects}
\alias{projects}
\alias{show_local_projects}
\alias{get_local_projects}
\alias{create_local_project}
\title{Manage projects}
\usage{
show_local_projects()

get_local_projects()

create_local_project(name)
}
\arguments{
\item{name}{Project name}
}
\description{
Manage projects
}
\section{Functions}{
\itemize{
\item \code{show_local_projects()}: Show the list of the existing projects

\item \code{get_local_projects()}: Get the list of the existing projects

\item \code{create_local_project()}: Create a new project board

}}
