% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{read-write}
\alias{read-write}
\alias{read_pin}
\alias{write_pin}
\title{Read/Write pins}
\usage{
read_pin(project = NULL, pin = NULL, ...)

write_pin(project, ...)
}
\arguments{
\item{project}{Project name}

\item{pin}{Pin name}

\item{...}{Other arguments passed to \code{pin_read()}/\code{pin_write()}}
}
\description{
Read/Write pins

Completeme version of the function
}
\section{Functions}{
\itemize{
\item \code{read_pin()}: Read pin

\item \code{write_pin()}: Write pin

}}
